Remove rl.envs.utils
