+++ tensor2tensor/data_generators/gym_problems.py
@@ -32,7 +32,6 @@ from tensor2tensor.data_generators import video_utils
 from tensor2tensor.models.research import rl
 from tensor2tensor.rl import collect
 from tensor2tensor.rl.envs import tf_atari_wrappers
-from tensor2tensor.rl.envs.utils import InitialFrameChooser
 from tensor2tensor.utils import metrics
 from tensor2tensor.utils import registry
